Updates the cause of build message for build configuration update
The possible causes of builds (as per current implementation of OpenShift based service) are:

1 -  First build of container (after indexing)
2 -  Git commit $SHA [..]
3 -  Upstream/parent project is rebuilt
4 -  Project configs (like desired-tag, emails) have updated
5 -  Manually triggered by admin

We have implemented 1, 2, 3 and 5.

For 4:
OpenShift / oc can identify only the new buildConfigs created -- meaning --> if buildConfig is updated for an existing project, OpenShift does not trigger the build by itself.

How we have handled this for builds?
A - We have implemented logic in our index reader. After reading index we identify, if a project configs are updated, we trigger the build using `oc start-build` command explicitly.

On the receiving end (i.e. notification module), when checks Jenkins job information, it sees, the build (whose configs are updated) is manually triggerred (since we ran oc `start-build` refer A)

Thus - programmaticaly, we have similar means for cases 4 and 5, i.e. When project configs are updated, the jenkins identifies it as manual trigger,
and there is no identifier for same in its current form. (given we are parsing Jenkins APIs to find cause of build)

I have kept the string "Manually triggerred by build." for case 5 (and implicitly 4).

For production scenario, I am sure, we are going to get case 4 very often as compare to case 5 (one of our team member manually starting the build).
Given this, do we want to change that message?

Keeping "Update to build configurations of container"